+++ Saper/Models/MinePlacement/IMinePlacementStrategy.cs
@@ -1,0 +1,7 @@
+namespace Saper.Models.MinePlacement
+{
+    public interface IMinePlacementStrategy
+    {
+        bool[] GenerateMines(int rows, int columns, int mineCount);
+    }
+}
